+++ packages/terra-framework-docs/src/terra-dev-site/doc/aggregator/common/placeholder-list/PlaceholderList.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import './PlaceholderList.module.scss';
 /* eslint-disable */
 const List = ({ children, isPadded }) => (
-  <ul aria-label="Placeholder List" role="listbox" className={isPadded ? 'placeholder-list is-padded' : 'placeholder-list '}>
+  <ul aria-label="Placeholder List" role="tablist" className={isPadded ? 'placeholder-list is-padded' : 'placeholder-list '}>
     {children}
   </ul>
 );