+++ packages/terra-table/CHANGELOG.md
@@ -2,13 +2,39 @@
 
 ## Unreleased
 
+* Added
+  * (flowsheet-data-grid only) Added support for subsections.
+
+* Changed
+  * Updated the cell component auto focus logic to be configured by the grid context.
+
+## 5.18.0 - (April 4, 2024)
+
+* Added
+  * Added the ability to configure the table to use auto table layout.
+
+* Fixed
+  * Fixed action header row so that it is properly fixed to the top of the table.
+
+## 5.17.0 - (March 29, 2024)
+
+* Added
+  * Added the ability to select a cell by pressing Escape on an auto focused button or anchor element (terra-data-grid only).
+
+## 5.16.0 - (March 25, 2024)
+
+* Fixed
+  * Fixed an issue where the click event for buttons inside table cells was not triggered by the spacebar.
+  
+## 5.15.0 - (March 21, 2024)
+
+* Changed
+  * Prevent right click from selecting row in row selection mode.
+
 ## 5.14.0 - (March 19, 2024)
 
 * Removed
   * `boundingRef` is no longer necessary for the section header's sticky behavior for tables. This change is still passive in nature as it doesn't change any functionality for consumers.
-
-* Changed
-  * Prevent right click from selecting row in row selection mode.
 
 ## 5.13.0 - (March 14, 2024)